Rename dbHandler.ts to realmHandler.ts
